% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandan_dat.R
\name{pandan_dat}
\alias{pandan_dat}
\title{Get project data}
\usage{
pandan_dat(project, gs_url = Sys.getenv("PANDAN_TRACKER"))
}
\arguments{
\item{project}{Project tracking sheet. First column for
components, second to some number of columns for project level progress,
and any column name containing \code{read_edit} will be classified as reading over
and editing, not project progress.}

\item{gs_url}{Url for googlesheet. Defaults to system environment variable
\code{PANDAN}. \code{usethis::edit_r_environ()} to set this.}
}
\value{
List of object of project metadata useful for visualisation, e.g.,
current levels and current sections.
}
\description{
Import project-level data from googlesheets.
}
\examples{
# runs because PANDAN url environment variable set and
# pandan googlesheet at url
project_dat("pandan")

}
