% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandan_gantt.R
\name{pandan_gantt}
\alias{pandan_gantt}
\title{Gantt chart}
\usage{
pandan_gantt(
  gantt_url = Sys.getenv("PANDAN_GANTT"),
  predecessor_sep = "\\\\s*;\\\\s*"
)
}
\arguments{
\item{gantt_url}{gs with sheets gantt and project key and meta}

\item{predecessor_sep}{Separator for predecessors, defaults to semicolon with any amount of white space around.}

\item{gantt_df}{Dataframe of the form project    task
predecessors    optimistic    normal    pessimistic
status added

where time is measured in days.}
}
\description{
A gantt chart is a type of bar chart with tasks performed in the y axis
and \strong{time} in the x axis. The horizontal width of the bar represents
a duration of time spent on a task within a \strong{project}.
}
\details{
The standard structure (according to wikipedia) is a data table of the form

task | predecessors |optimistic (O)| normal (M) | pessimistic (P) | expected_time (E)

Predecessors can can contain multiple values, so will need an argument for
separator.

The expected time is estimated using a Beta distribution statistic,

$$
(O + 4M + P) / 6
$$
}
