% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandan_instantiate.R
\name{pandan_instantiate}
\alias{pandan_instantiate}
\title{Instantiate a project}
\usage{
pandan_instantiate(
  project,
  objective,
  category,
  group,
  gs_url = Sys.getenv("PANDAN_TRACKER")
)
}
\arguments{
\item{project}{Project tracking sheet. First column for
components, second to some number of columns for project level progress,
and any column name containing \code{read_edit} will be classified as reading over
and editing, not project progress.}

\item{objective}{A string describing what finished looks like.}

\item{category}{Specify talk or manuscript.}

\item{group}{Specify meta category}

\item{gs_url}{Url for googlesheet. Defaults to system environment variable
\code{PANDAN}. \code{usethis::edit_r_environ()} to set this.}
}
\description{
Copy the project template to sheet.
}
\details{
Add project name, description, and url to \href{https://docs.google.com/spreadsheets/d/1q1dJMjJRVgH6nFW_59RngF4YBkHX1RW2ZgekZVIE_zY/edit#gid=0}{pandan projects}.

Add 0 progress to pandan progress log.
}
