% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandan_next.R
\name{pandan_next}
\alias{pandan_next}
\title{Get next actions}
\usage{
pandan_next(
  project,
  gs_url = Sys.getenv("PANDAN_TRACKER"),
  gs_projects = Sys.getenv("PANDAN_PROJECTS")
)
}
\arguments{
\item{project}{Project tracking sheet. First column for
components, second to some number of columns for project level progress,
and any column name containing \code{read_edit} will be classified as reading over
and editing, not project progress.}

\item{gs_url}{Url for googlesheet. Defaults to system environment variable
\code{PANDAN}. \code{usethis::edit_r_environ()} to set this.}
}
\description{
Next three things to write and where to read from.
}
