% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandan_tracker.R
\name{pandan_tracker}
\alias{pandan_tracker}
\title{Open a pandan sheet}
\usage{
pandan_tracker(gs_url = Sys.getenv("PANDAN_MS"))
}
\arguments{
\item{gs_url}{Url for googlesheet. Defaults to system environment variable
\code{PANDAN}. \code{usethis::edit_r_environ()} to set this.}
}
\description{
Open a pandan sheet
}
