% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandan_view.R
\name{pandan_view}
\alias{pandan_view}
\title{See progress over time}
\usage{
pandan_view(project = "all", distill = FALSE)
}
\arguments{
\item{distill}{Plot theme for matching distill blog.}
}
\description{
See progress over time
}
