% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandan_gantt.R
\name{wrangle_gantt_dat}
\alias{wrangle_gantt_dat}
\title{Wrangle data for Gantt plot}
\usage{
wrangle_gantt_dat(
  gantt_df = googlesheets4::read_sheet(Sys.getenv("PANDAN_GANTT"), "gantt"),
  project_key = googlesheets4::read_sheet(Sys.getenv("PANDAN_GANTT"), "project_key"),
  meta_df = googlesheets4::read_sheet(Sys.getenv("PANDAN_GANTT"), "meta"),
  predecessor_sep = "\\\\s*;\\\\s*"
)
}
\arguments{
\item{predecessor_sep}{Separator for predecessors, defaults to semicolon
with any amount of white space around.}
}
\description{
Wrangle data for Gantt plot
}
