% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandan_tracker.R
\name{pandan_tracker}
\alias{pandan_tracker}
\title{Open a pandan sheet}
\usage{
pandan_tracker(tracker = "project")
}
\arguments{
\item{tracker}{Specify tracker, progress, or projects to open sheet.}
}
\description{
Open a pandan sheet
}
